+++ nemo_skills/training/openrlhf/math_reward.py
@@ -17,7 +17,7 @@ import os
 
 import torch
 
-from nemo_skills.code_execution.math_grader import extract_answer
+from nemo_skills.evaluation.math_grader import extract_answer
 from nemo_skills.evaluation.metrics.utils import is_correct_judgement
 from nemo_skills.inference.server.model import get_model
 from nemo_skills.prompt.utils import get_prompt
